[client] moved getting event's label in converter base class
