+++ client/callgrind_converter.cpp
@@ -1,17 +1,11 @@
 #include "callgrind_converter.hpp"
 
-#include "hawktracer/parser/klass_register.hpp"
 #include <algorithm>
 
 namespace HawkTracer
 {
 namespace client
 {
-
-CallgrindConverter::CallgrindConverter() :
-    _mapping_klass_name("HT_StringMappingEvent")
-{
-}
 
 CallgrindConverter::~CallgrindConverter()
 {
@@ -120,46 +114,7 @@ void CallgrindConverter::_add_event(uint32_t thread_id,
 
 void CallgrindConverter::process_event(const parser::Event& event)
 {
-    std::string label;
-
-    if (_mapping_klass_id == 0 &&
-            event.get_klass()->get_id() == HawkTracer::parser::to_underlying(HawkTracer::parser::WellKnownKlasses::EventKlassInfoEventKlass))
-    {
-        if (event.get_value<char*>("event_klass_name") == _mapping_klass_name)
-        {
-            _mapping_klass_id = event.get_value<HT_EventKlassId>("info_klass_id");
-        }
-    }
-    if (event.get_klass()->get_id() == _mapping_klass_id)
-    {
-        _tracepoint_map->add_map_entry(event.get_value<uint64_t>("identifier"), event.get_value<char*>("label"));
-        return;
-    }
-
-    if (event.has_value("label"))
-    {
-        const parser::Event::Value& value = event.get_raw_value("label");
-        if (value.field->get_type_id() == parser::FieldTypeId::UINT64)
-        {
-            label = _tracepoint_map->get_label_info(value.value.f_UINT64).label;
-        }
-        else if (value.field->get_type_id() == parser::FieldTypeId::STRING)
-        {
-            label = value.value.f_STRING;
-        }
-        else
-        {
-            label = "invalid label type";
-        }
-    }
-    else if (event.has_value("name"))
-    {
-        label = event.get_value<char*>("name");
-    }
-    else
-    {
-        return;
-    }
+    std::string label = get_label(event);
 
     uint32_t thread_id = event.get_value_or_default<uint32_t>("thread_id", 0);
     HT_TimestampNs start_ts = event.get_value<uint64_t>("timestamp");
